+++ aioxmpp/stream.py
@@ -927,7 +927,33 @@ class StanzaStream:
             self.on_stream_destroyed(exc)
             self._established = False
 
-    def _iq_request_coro_done(self, request, task):
+    def _compose_undefined_condition(self, request):
+        response = request.make_reply(type_=structs.IQType.ERROR)
+        response.error = stanza.Error(
+            condition=errors.ErrorCondition.UNDEFINED_CONDITION,
+            type_=structs.ErrorType.CANCEL,
+        )
+        return response
+
+    def _send_iq_reply(self, request, result):
+        try:
+            if isinstance(result, errors.XMPPError):
+                response = request.make_reply(type_=structs.IQType.ERROR)
+                response.error = stanza.Error.from_exception(result)
+            else:
+                response = request.make_reply(type_=structs.IQType.RESULT)
+                response.payload = result
+        except Exception:
+            self._logger.exception("invalid payload for an IQ response")
+            response = self._compose_undefined_condition(
+                request
+            )
+        self._enqueue(response)
+
+    def _iq_request_coro_done_remove_task(self, task):
+        self._iq_request_tasks.remove(task)
+
+    def _iq_request_coro_done_send_reply(self, request, task):
         """
         Called when an IQ request handler coroutine returns. `request` holds
         the IQ request which triggered the excecution of the coroutine and
@@ -935,23 +961,22 @@ class StanzaStream:
 
         Compose a response and send that response.
         """
-        self._iq_request_tasks.remove(task)
         try:
             payload = task.result()
         except errors.XMPPError as err:
-            response = request.make_reply(type_=structs.IQType.ERROR)
-            response.error = stanza.Error.from_exception(err)
+            self._send_iq_reply(request, err)
         except Exception:
-            response = request.make_reply(type_=structs.IQType.ERROR)
-            response.error = stanza.Error(
-                condition=errors.ErrorCondition.UNDEFINED_CONDITION,
-                type_=structs.ErrorType.CANCEL,
-            )
+            response = self._compose_undefined_condition(request)
+            self._enqueue(response)
             self._logger.exception("IQ request coroutine failed")
         else:
-            response = request.make_reply(type_=structs.IQType.RESULT)
-            response.payload = payload
-        self._enqueue(response)
+            self._send_iq_reply(request, payload)
+
+    def _iq_request_coro_done_check(self, task):
+        try:
+            task.result()
+        except Exception:
+            self._logger.exception("IQ request coroutine failed")
 
     def _process_incoming_iq(self, stanza_obj):
         """
@@ -986,7 +1011,7 @@ class StanzaStream:
             self._logger.debug("iq is request")
             key = (stanza_obj.type_, type(stanza_obj.payload))
             try:
-                coro = self._iq_request_map[key]
+                coro, with_send_reply = self._iq_request_map[key]
             except KeyError:
                 self._logger.warning(
                     "unhandleable IQ request: from=%r, type_=%r, payload=%r",
@@ -1001,17 +1026,34 @@ class StanzaStream:
                 self._enqueue(response)
                 return
 
+            args = [stanza_obj]
+            if with_send_reply:
+
+                def send_reply(result=None):
+                    nonlocal task, stanza_obj, send_reply_callback
+                    if task.done():
+                        raise RuntimeError(
+                            "send_reply called after the handler is done")
+                    if task.remove_done_callback(send_reply_callback) == 0:
+                        raise RuntimeError(
+                            "send_reply called more than once")
+                    task.add_done_callback(self._iq_request_coro_done_check)
+                    self._send_iq_reply(stanza_obj, result)
+
+                args.append(send_reply)
+
             try:
-                awaitable = coro(stanza_obj)
+                awaitable = coro(*args)
             except Exception as exc:
                 awaitable = asyncio.Future()
                 awaitable.set_exception(exc)
 
             task = asyncio.ensure_future(awaitable)
-            task.add_done_callback(
-                functools.partial(
-                    self._iq_request_coro_done,
-                    stanza_obj))
+            send_reply_callback = functools.partial(
+                self._iq_request_coro_done_send_reply,
+                stanza_obj)
+            task.add_done_callback(self._iq_request_coro_done_remove_task)
+            task.add_done_callback(send_reply_callback)
             self._iq_request_tasks.append(task)
             self._logger.debug("started task to handle request: %r", task)
 
@@ -1352,7 +1394,8 @@ class StanzaStream:
             stacklevel=2)
         return self.register_iq_request_handler(type_, payload_cls, coro)
 
-    def register_iq_request_handler(self, type_, payload_cls, cb):
+    def register_iq_request_handler(self, type_, payload_cls, cb, *,
+                                    with_send_reply=False):
         """
         Register a coroutine function or a function returning an awaitable to
         run when an IQ request is received.
@@ -1363,6 +1406,9 @@ class StanzaStream:
             :class:`~xso.XSO`)
         :type payload_cls: :class:`~.XMLStreamClass`
         :param cb: Function or coroutine function to invoke
+        :param with_send_reply: Whether to pass a function to send a reply
+             to `cb` as second argument.
+        :type with_send_reply: :class:`bool`
         :raises ValueError: if there is already a coroutine registered for this
                             target
         :raises ValueError: if `type_` is not a request IQ type
@@ -1407,7 +1453,21 @@ class StanzaStream:
 
             Using a non-coroutine function for `cb` will generally lead to
             less readable code. For the sake of readability, it is recommended
-            prefer coroutine functions.
+            to prefer coroutine functions when strong ordering guarantees are
+            not needed.
+
+        .. versionadded:: 0.11
+
+            When the argument `with_send_reply` is true `cb` will be
+            called with two arguments: the IQ stanza to handle and a
+            unary function `send_reply(result=None)` that sends a
+            response to the IQ request and prevents that an automatic
+            response is sent. If `result` is an instance of
+            :class:`~aioxmpp.XMPPError` an error result is generated.
+
+            This is useful when the handler function needs to execute
+            actions which happen after the IQ result has been sent,
+            for example, sending other stanzas.
 
         .. versionchanged:: 0.10
 
@@ -1448,7 +1508,7 @@ class StanzaStream:
         if key in self._iq_request_map:
             raise ValueError("only one listener is allowed per tag")
 
-        self._iq_request_map[key] = cb
+        self._iq_request_map[key] = cb, with_send_reply
         self._logger.debug(
             "iq request coroutine registered: type=%r, payload=%r",
             type_, payload_cls)
@@ -2620,7 +2680,7 @@ class StanzaStream:
 
 
 @contextlib.contextmanager
-def iq_handler(stream, type_, payload_cls, coro):
+def iq_handler(stream, type_, payload_cls, coro, *, with_send_reply=False):
     """
     Context manager to temporarily register a coroutine to handle IQ requests
     on a :class:`StanzaStream`.
@@ -2633,10 +2693,19 @@ def iq_handler(stream, type_, payload_cls, coro):
                         :class:`~xso.XSO`)
     :type payload_cls: :class:`~.XMLStreamClass`
     :param coro: Coroutine to register
+    :param with_send_reply: Whether to pass a function to send the reply
+                            early to `cb`.
+    :type with_send_reply: :class:`bool`
 
     The coroutine is registered when the context is entered and unregistered
     when the context is exited. Running coroutines are not affected by exiting
     the context manager.
+
+    .. versionadded:: 0.11
+
+       The `with_send_reply` argument. See
+       :meth:`aioxmpp.stream.StanzaStream.register_iq_request_handler` for
+       more detail.
 
     .. versionadded:: 0.8
     """
@@ -2645,6 +2714,7 @@ def iq_handler(stream, type_, payload_cls, coro):
         type_,
         payload_cls,
         coro,
+        with_send_reply=with_send_reply,
     )
     try:
         yield